Moved ActivityPub service to core package
fix https://linear.app/ghost/issue/ENG-2063/activitypub

- this moves the AP service and test to the core package now we support
  TS in there
- this allows us to start paring down the code we have, and relying on
  TS more
- we can now also remove the activitypub package as it's no longer needed
