+++ ghost/core/package.json
@@ -62,7 +62,6 @@
     "@extractus/oembed-extractor": "3.2.1",
     "@sentry/node": "7.119.2",
     "@slack/webhook": "7.0.4",
-    "@tryghost/activitypub": "0.0.0",
     "@tryghost/adapter-base-cache": "0.1.12",
     "@tryghost/adapter-cache-redis": "0.0.0",
     "@tryghost/adapter-manager": "0.0.0",
@@ -219,6 +218,7 @@
     "multer": "1.4.4",
     "mysql2": "3.12.0",
     "nconf": "0.12.1",
+    "node-fetch": "2.7.0",
     "node-jose": "2.2.0",
     "path-match": "1.2.4",
     "probe-image-size": "7.2.3",
